+++ packages/taro-h5/src/api/media/background-audio/index.ts
@@ -1,4 +1,5 @@
-import { temporarilyNotSupport } from '../utils'
+import { temporarilyNotSupport } from '../../utils'
+import { BackgroundAudioManager } from './backgroundAudioManager'
 
 // 背景音频
 export const stopBackgroundAudio = temporarilyNotSupport('stopBackgroundAudio')
@@ -9,4 +10,8 @@ export const onBackgroundAudioStop = temporarilyNotSupport('onBackgroundAudioSto
 export const onBackgroundAudioPlay = temporarilyNotSupport('onBackgroundAudioPlay')
 export const onBackgroundAudioPause = temporarilyNotSupport('onBackgroundAudioPause')
 export const getBackgroundAudioPlayerState = temporarilyNotSupport('getBackgroundAudioPlayerState')
-export const getBackgroundAudioManager = temporarilyNotSupport('getBackgroundAudioManager')
+
+/**
+ * 获取全局唯一的背景音频管理器
+ */
+export const getBackgroundAudioManager = () => new BackgroundAudioManager()
